fixed Babel Problem #261
before this fix, all translations function are related to `Domain` of
flask-admin, which can only use the .mo file shipped with flask-admin.
And this fix make all quokka scoped translation function to use the
default domain of flask-babelex, which means it will find the *.mo files
in quokka/quokka/translations directory.
It will rely the speaklater package to accept PR#6.